+++ src/american-express/index.js
@@ -9,7 +9,7 @@ var AmericanExpress = require('./american-express');
 var sharedErrors = require('../lib/errors');
 var VERSION = process.env.npm_package_version;
 var Promise = require('../lib/promise');
-var wrapPromise = require('wrap-promise');
+var wrapPromise = require('@braintree/wrap-promise');
 
 /**
  * @static