+++ src/unionpay/index.js
@@ -11,7 +11,7 @@ var errors = require('./shared/errors');
 var sharedErrors = require('../lib/errors');
 var VERSION = process.env.npm_package_version;
 var Promise = require('../lib/promise');
-var wrapPromise = require('wrap-promise');
+var wrapPromise = require('@braintree/wrap-promise');
 
 /**
 * @static