+++ src/us-bank-account/us-bank-account.js
@@ -10,7 +10,7 @@ var convertMethodsToError = require('../lib/convert-methods-to-error');
 var methods = require('../lib/methods');
 var camelCaseToSnakeCase = require('../lib/camel-case-to-snake-case');
 var Promise = require('../lib/promise');
-var wrapPromise = require('wrap-promise');
+var wrapPromise = require('@braintree/wrap-promise');
 
 /**
  * @typedef {object} USBankAccount~tokenizePayload